+++ Plugins/Sentry/Source/Sentry/Private/Windows/WindowsSentrySubsystem.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#if USE_SENTRY_NATIVE
+#if USE_SENTRY_NATIVE && !SENTRY_WINGDK
 
 #include "Utils/SentryPlatformInfo.h"
 
@@ -17,6 +17,7 @@ protected:
 	virtual void ConfigureHandlerPath(sentry_options_t* Options) override;
 	virtual void ConfigureStackCaptureStrategy(sentry_options_t* Options) override;
 	virtual void ConfigureCrashReporterPath(sentry_options_t* Options) override;
+	virtual void ConfigureScreenshotCapturing(sentry_options_t* Options) override;
 
 	virtual FString GetHandlerExecutableName() const override;
 	virtual FString GetCrashReporterExecutableName() const override { return TEXT("Sentry.CrashReporter.exe"); }
@@ -24,6 +25,7 @@ protected:
 	virtual sentry_value_t OnCrash(const sentry_ucontext_t* uctx, sentry_value_t event, void* closure) override;
 
 	virtual bool IsScreenshotSupported() const override { return true; }
+	virtual bool IsOutOfProcessScreenshotEnabled() const override { return bOutOfProcessScreenshots; }
 	virtual bool IsHangTrackingSupported() const override { return true; }
 
 	virtual FString GetDeviceType() const override;
@@ -31,8 +33,11 @@ protected:
 private:
 	/** Wine/Proton detection info */
 	FWineProtonInfo WineProtonInfo;
+
+	/** Whether native out-of-process screenshot capturing is enabled */
+	bool bOutOfProcessScreenshots = false;
 };
 
 typedef FWindowsSentrySubsystem FPlatformSentrySubsystem;
 
-#endif // USE_SENTRY_NATIVE
+#endif // USE_SENTRY_NATIVE && !SENTRY_WINGDK